Add Night Soldier
Sorry for giant commit. This includes the general EnemyAI class, which *should* be enough for all enemy types I think. I've made a prefab for Night Soldier and some very very basic shooting for now. I've also refactored some of my other code, added health script to the Player prefab and changed Player's tag to Player because I'm using it in my code to check if the object is a player.